+++ lib/split_view.dart
@@ -10,6 +10,10 @@ class SplitView extends StatefulWidget {
   final Widget view2;
   final SplitViewMode viewMode;
   final double gripSize;
+  final double? minWidthSidebar;
+  final double? maxWidthSidebar;
+  final double? minHeightSidebar;
+  final double? maxHeightSidebar;
   final double initialWeight;
   final Color gripColor;
   final double positionLimit;
@@ -21,6 +25,10 @@ class SplitView extends StatefulWidget {
     required this.view2,
     required this.viewMode,
     this.gripSize = 12.0,
+    this.minWidthSidebar,
+    this.maxWidthSidebar,
+    this.minHeightSidebar,
+    this.maxHeightSidebar,
     this.initialWeight = 0.5,
     this.gripColor = Colors.grey,
     this.positionLimit = 20.0,
@@ -41,15 +49,12 @@ class _SplitViewState extends State<SplitView> {
     super.didChangeDependencies();
 
     this.defaultWeight =
-        PageStorage.of(context)?.readState(context, identifier: widget.key) ??
-            widget.initialWeight;
+        PageStorage.of(context)?.readState(context, identifier: widget.key) ?? widget.initialWeight;
+    weight = ValueNotifier(defaultWeight);
   }
 
   @override
   Widget build(BuildContext context) {
-    weight = ValueNotifier(defaultWeight);
-    _prevWeight = defaultWeight;
-
     return LayoutBuilder(
       builder: (context, constraints) {
         return ValueListenableBuilder<double?>(
@@ -57,8 +62,7 @@ class _SplitViewState extends State<SplitView> {
           builder: (_, w, __) {
             if (widget.onWeightChanged != null && _prevWeight != w) {
               _prevWeight = w;
-              PageStorage.of(context)
-                  ?.writeState(context, w, identifier: widget.key);
+              PageStorage.of(context)?.writeState(context, w, identifier: widget.key);
               widget.onWeightChanged!(w);
             }
             if (widget.viewMode == SplitViewMode.Vertical) {
@@ -72,11 +76,18 @@ class _SplitViewState extends State<SplitView> {
     );
   }
 
-  Stack _buildVerticalView(
-      BuildContext context, BoxConstraints constraints, double w) {
-    final double top = constraints.maxHeight * w;
-    final double bottom = constraints.maxHeight * (1.0 - w);
+  Stack _buildVerticalView(BuildContext context, BoxConstraints constraints, double w) {
+    double top = constraints.maxHeight * w;
+    double bottom = constraints.maxHeight * (1.0 - w);
     final halfGripSize = widget.gripSize / 2.0;
+
+    if (widget.maxHeightSidebar != null && top > widget.maxHeightSidebar!) {
+      top = widget.maxHeightSidebar!;
+      bottom = constraints.maxHeight - widget.maxHeightSidebar!;
+    } else if (widget.minHeightSidebar != null && top < widget.minHeightSidebar!) {
+      top = widget.minHeightSidebar!;
+      bottom = constraints.maxHeight - widget.minHeightSidebar!;
+    }
 
     return Stack(
       children: <Widget>[
@@ -104,8 +115,7 @@ class _SplitViewState extends State<SplitView> {
             child: GestureDetector(
               behavior: HitTestBehavior.translucent,
               onVerticalDragUpdate: (detail) {
-                final RenderBox container =
-                    context.findRenderObject() as RenderBox;
+                final RenderBox container = context.findRenderObject() as RenderBox;
                 final pos = container.globalToLocal(detail.globalPosition);
                 if (pos.dy > widget.positionLimit &&
                     pos.dy < (container.size.height - widget.positionLimit)) {
@@ -120,11 +130,18 @@ class _SplitViewState extends State<SplitView> {
     );
   }
 
-  Widget _buildHorizontalView(
-      BuildContext context, BoxConstraints constraints, double w) {
-    final double left = constraints.maxWidth * w;
-    final double right = constraints.maxWidth * (1.0 - w);
+  Widget _buildHorizontalView(BuildContext context, BoxConstraints constraints, double w) {
+    double left = constraints.maxWidth * w;
+    double right = constraints.maxWidth * (1.0 - w);
     final double halfGripSize = widget.gripSize / 2.0;
+
+    if (widget.maxWidthSidebar != null && left > widget.maxWidthSidebar!) {
+      left = widget.maxWidthSidebar!;
+      right = constraints.maxWidth - widget.maxWidthSidebar!;
+    } else if (widget.minWidthSidebar != null && left < widget.minWidthSidebar!) {
+      left = widget.minWidthSidebar!;
+      right = constraints.maxWidth - widget.minWidthSidebar!;
+    }
 
     return Stack(
       children: <Widget>[
@@ -152,8 +169,7 @@ class _SplitViewState extends State<SplitView> {
             child: GestureDetector(
               behavior: HitTestBehavior.translucent,
               onVerticalDragUpdate: (detail) {
-                final RenderBox container =
-                    context.findRenderObject() as RenderBox;
+                final RenderBox container = context.findRenderObject() as RenderBox;
                 final pos = container.globalToLocal(detail.globalPosition);
                 if (pos.dx > widget.positionLimit &&
                     pos.dx < (container.size.width - widget.positionLimit)) {